requirements.txt
asgiref==3.3.4
Django==3.2.7
djangorestframework==3.12.4
drf-jwt==1.19.1
pytz==2022.4
sqlparse==0.4.1
redis==3.5.3
celery==5.2.6
psycopg2-binary==2.8.6
flower==1.2.0
requests==2.25.1
xmltodict==0.12.0
flake8==3.9.2
django-cors-headers==3.7.0
django-utils-six==2.0
drf-yasg==1.21.4
